Unison.ABT

Documentation

data ABT f v r

Constructors

Var v

Cycle r

Abs v r

Tm (f r)

data Term f v a

At each level in the tree, we store the set of free variables and a value of type a. Variables are of type v.

Constructors

Term

Fields

freeVars :: Set v

annotation :: a

out :: ABT f v (Term f v a)

class Ord v => Var v where

A class for variables.

`Set.notMember (freshIn vs v) vs`: freshIn returns a variable not used in the Set

Methods

freshIn :: Set v -> v -> v

data V v

Constructors

Free v

Bound v

unvar :: V v -> v

newtype Path s t a b m

Constructors

Path

Fields

focus :: s -> Maybe (a, b -> Maybe t, m)

here :: Monoid m => Path s t s t m

type Path' f g m = forall a v. Var v => Path (Term f v a) (Term f (V v) a) (Term g v a) (Term g (V v) a) m

compose :: Monoid m => Path s t a b m -> Path a b a' b' m -> Path s t a' b' m

at :: Path s t a b m -> s -> Maybe a

modify' :: Path s t a b m -> (m -> a -> b) -> s -> Maybe t

wrap :: (Functor f, Foldable f, Var v) => v -> Term f (V v) a -> (V v, Term f (V v) a)

wrap' :: (Functor f, Foldable f, Var v) => v -> Term f (V v) a -> (V v -> Term f (V v) a -> c) -> c

bound' :: Foldable f => Term f v a -> [v]

Return the list of all variables bound by this ABT

annotateBound' :: (Ord v, Functor f, Foldable f) => Term f v a0 -> Term f v [v]

annotateBound :: (Ord v, Foldable f, Functor f) => Term f v a -> Term f v (a, Set v)

annotateBound'' :: (Ord v, Functor f, Foldable f) => Term f v a -> Term f v (a, [v])

bound :: (Ord v, Foldable f) => Term f v a -> Set v

Return the set of all variables bound by this ABT

isClosed :: Term f v a -> Bool

True if the term has no free variables, False otherwise

isFreeIn :: Ord v => v -> Term f v a -> Bool

True if v is a member of the set of free variables of t

annotate :: a -> Term f v a -> Term f v a

Replace the annotation with the given argument.

vmap :: (Functor f, Foldable f, Ord v2) => (v -> v2) -> Term f v a -> Term f v2 a

vmapM :: (Applicative m, Traversable f, Foldable f, Ord v2) => (v -> m v2) -> Term f v a -> m (Term f v2 a)

amap :: (Functor f, Foldable f, Ord v) => (a -> a2) -> Term f v a -> Term f v a2

amap' :: (Functor f, Foldable f, Ord v) => (Term f v a -> a -> a2) -> Term f v a -> Term f v a2

extraMap :: Functor g => (forall k. f k -> g k) -> Term f v a -> Term g v a

pattern Var' :: forall f v a. v -> Term f v a

pattern Cycle' :: forall f v a. [v] -> Term f v a -> Term f v a

pattern Abs' :: (Foldable f, Functor f, Var v) => Subst f v a -> Term f v a

pattern AbsN' :: forall f v a. [v] -> Term f v a -> Term f v a

pattern Tm' :: f (Term f v a) -> Term f v a

pattern CycleA' :: forall f v a. a -> [(a, v)] -> Term f v a -> Term f v a

pattern AbsNA' :: forall f v a. [(a, v)] -> Term f v a -> Term f v a

pattern Abs1NA' :: forall f v a. [(a, v)] -> Term f v a -> Term f v a

unabsA :: Term f v a -> ([(a, v)], Term f v a)

unabs1A :: Term f v a -> Maybe ([(a, v)], Term f v a)

var :: v -> Term f v ()

annotatedVar :: a -> v -> Term f v a

abs :: Ord v => v -> Term f v () -> Term f v ()

abs' :: Ord v => a -> v -> Term f v a -> Term f v a

absr :: (Functor f, Foldable f, Var v) => v -> Term f (V v) () -> Term f (V v) ()

absr' :: (Functor f, Foldable f, Var v) => a -> v -> Term f (V v) a -> Term f (V v) a

Rebuild an abs, renaming v to avoid capturing any `Free v` in body.

absChain :: Ord v => [v] -> Term f v () -> Term f v ()

absCycle :: Ord v => [v] -> Term f v () -> Term f v ()

absChain' :: Ord v => [(a, v)] -> Term f v a -> Term f v a

tm :: (Foldable f, Ord v) => f (Term f v ()) -> Term f v ()

tm' :: (Foldable f, Ord v) => a -> f (Term f v a) -> Term f v a

cycle :: Term f v () -> Term f v ()

cycle' :: a -> Term f v a -> Term f v a

cycler' :: (Functor f, Foldable f, Var v) => a -> [v] -> Term f (V v) a -> Term f (V v) a

cycler :: (Functor f, Foldable f, Var v) => [v] -> Term f (V v) () -> Term f (V v) ()

into :: (Foldable f, Ord v) => ABT f v (Term f v ()) -> Term f v ()

into' :: (Foldable f, Ord v) => a -> ABT f v (Term f v a) -> Term f v a

rename :: (Foldable f, Functor f, Var v) => v -> v -> Term f v a -> Term f v a

renames old to new in the given term, ignoring subtrees that bind old

renames :: (Foldable f, Functor f, Var v) => Map v v -> Term f v a -> Term f v a

changeVars :: (Foldable f, Functor f, Var v) => Map v v -> Term f v a -> Term f v a

freshInBoth :: Var v => Term f v a -> Term f v a -> v -> v

Produce a variable which is free in both terms

fresh :: Var v => Term f v a -> v -> v

freshEverywhere :: (Foldable f, Var v) => Term f v a -> v -> v

allVars :: Foldable f => Term f v a -> [v]

freshes :: Var v => Term f v a -> [v] -> [v]

freshes' :: Var v => Set v -> [v] -> [v]

freshenS :: (Var v, MonadState (Set v) m) => v -> m v

Freshens the given variable wrt. the set of used variables tracked by state. Adds the result to the set of used variables.

freshenS' :: (Var v, MonadState s m) => Lens' s (Set v) -> v -> m v

A more general version of freshenS that uses a lens to focus on used variables inside state.

subst :: (Foldable f, Functor f, Var v) => v -> Term f v a -> Term f v a -> Term f v a

`subst v e body` substitutes e for v in body, avoiding capture by renaming abstractions in body

subst' :: (Foldable f, Functor f, Var v) => (a -> Term f v a) -> v -> Set v -> Term f v a -> Term f v a

substInheritAnnotation :: (Foldable f, Functor f, Var v) => v -> Term f v b -> Term f v a -> Term f v a

substsInheritAnnotation :: (Foldable f, Functor f, Var v) => [(v, Term f v b)] -> Term f v a -> Term f v a

substs :: (Foldable f, Functor f, Var v) => [(v, Term f v a)] -> Term f v a -> Term f v a

`substs [(t1,v1), (t2,v2), ...] body` performs multiple simultaneous substitutions, avoiding capture

occurrences :: (Foldable f, Var v) => v -> Term f v a -> Int

rebuildUp :: (Ord v, Foldable f, Functor f) => (f (Term f v a) -> f (Term f v a)) -> Term f v a -> Term f v a

rebuildUp' :: (Ord v, Foldable f, Functor f) => (Term f v a -> Term f v a) -> Term f v a -> Term f v a

freeVarOccurrences :: (Traversable f, Ord v) => Set v -> Term f v a -> [(v, a)]

foreachSubterm :: (Traversable f, Applicative g, Ord v) => (Term f v a -> g b) -> Term f v a -> g [b]

subterms :: (Ord v, Traversable f) => Term f v a -> [Term f v a]

visit :: (Traversable f, Applicative g, Ord v) => (Term f v a -> Maybe (g (Term f v a))) -> Term f v a -> g (Term f v a)

`visit f t` applies an effectful function to each subtree of t and sequences the results. When f returns Nothing, visit descends into the children of the current subtree. When f returns `Just t2`, visit replaces the current subtree with t2. Thus: `visit (const Nothing) t == pure t` and `visit (const (Just (pure t2))) t == pure t2`

visit' :: (Traversable f, Applicative g, Monad g, Ord v) => (f (Term f v a) -> g (f (Term f v a))) -> Term f v a -> g (Term f v a)

Apply an effectful function to an ABT tree top down, sequencing the results.

visitPure :: (Traversable f, Ord v) => (Term f v a -> Maybe (Term f v a)) -> Term f v a -> Term f v a

visit specialized to the Identity effect.

rewriteDown :: (Traversable f, Ord v) => (Term f v a -> Term f v a) -> Term f v a -> Term f v a

data Subst f v a

Constructors

Subst

Fields

freshen :: forall m v'. Monad m => (v -> m v') -> m v'

bind :: Term f v a -> Term f v a

bindInheritAnnotation :: forall b. Term f v b -> Term f v a

variable :: v

unabs1 :: (Foldable f, Functor f, Var v) => Term f v a -> Maybe (Subst f v a)

unabs :: Term f v a -> ([v], Term f v a)

reabs :: Ord v => [v] -> Term f v () -> Term f v ()

transform :: (Ord v, Foldable g, Functor f) => (forall a. f a -> g a) -> Term f v a -> Term g v a

transformM :: (Ord v, Monad m, Traversable g) => (forall a. f a -> m (g a)) -> Term f v a -> m (Term g v a)

reannotateUp :: (Ord v, Foldable f, Functor f, Monoid b) => (Term f v a -> b) -> Term f v a -> Term f v (a, b)

data FindAction x

Constructors

Found x

Prune

Continue

find :: (Ord v, Foldable f, Functor f) => (Term f v a -> FindAction x) -> Term f v a -> [x]

find' :: (Ord v, Foldable f, Functor f) => (Term f v a -> Bool) -> Term f v a -> [Term f v a]

components :: Var v => [(v, Term f v a)] -> [[(v, Term f v a)]]

orderedComponents' :: Var v => [(v, Term f v a)] -> [[(v, Term f v a)]]

orderedComponents :: Var v => [(v, Term f v a)] -> [[(v, Term f v a)]]

hashComponent :: (Functor f, Hashable1 f, Foldable f, Eq v, Show v, Ord v, Ord h, Accumulate h) => Map v (Term f v a) -> (h, [(v, Term f v a)])

hashComponents :: (Functor f, Hashable1 f, Foldable f, Eq v, Show v, Var v, Ord h, Accumulate h) => (h -> Word64 -> Word64 -> Term f v ()) -> Map v (Term f v a) -> [(h, [(v, Term f v a)])]

data Component f a

Constructors

Component [a] a

Embed (f a)

hash :: forall f v a h. (Functor f, Hashable1 f, Eq v, Show v, Ord h, Accumulate h) => Term f v a -> h

We ignore annotations in the Term, as these should never affect the meaning of the term.

distinct :: forall f v h a proxy. (Functor f, Hashable1 f, Eq v, Show v, Ord h, Accumulate h) => proxy h -> [Term f v a] -> [Term f v a]

Use the hash function to efficiently remove duplicates from the list, preserving order.

subtract :: forall f v h a proxy. (Functor f, Hashable1 f, Eq v, Show v, Ord h, Accumulate h) => proxy h -> [Term f v a] -> [Term f v a] -> [Term f v a]

Use the hash function to remove elements from t1s that exist in t2s, preserving order.