Unison.Blank

Documentation

data Blank loc

Blank is just a dummy annotation.

Constructors

Blank
just a dummy annotation

Recorded (Recorded loc)
indicates that we want to remember the variable's solution for some reason

Retain
indicates that we want to prefer keeping the variable in the context to better refine the above recorded solutions

data Recorded loc

Constructors

Placeholder loc String

Resolve loc String

MissingResultPlaceholder loc

loc :: Recorded loc -> loc

nameb :: Blank loc -> Maybe String