Unison.ConstructorReference

Description

The constructor reference type.

Documentation

data GConstructorReference r

A reference to a constructor is represented by a reference to its type declaration, plus the ordinal constructor id.

Constructors

ConstructorReference !r !ConstructorId

type ConstructorReference = GConstructorReference TypeReference

type ConstructorReferenceId = GConstructorReference TypeReferenceId

reference_ :: forall r s f. Functor f => (r -> f s) -> GConstructorReference r -> f (GConstructorReference s)

A lens onto the reference part of a constructor reference.

toId :: ConstructorReference -> Maybe ConstructorReferenceId

toShortHash :: ConstructorReference -> ShortHash

toText :: ConstructorReference -> Text