Unison.Name

Documentation

data Name

A name is an absolute-or-relative non-empty list of name segments.

class Convert a b where

Methods

convert :: a -> b

class Parse a b where

Methods

parse :: a -> Maybe b

Basic construction

cons :: HasCallStack => NameSegment -> Name -> Name

Cons a name segment onto the head of a relative name. Not monotonic with respect to ordering! It is not safe to use cons s as the first argument to Map.mapKeysMonotonic!

Precondition: the name is relative

O(n), where n is the number of segments.

joinDot :: HasCallStack => Name -> Name -> Name

fromSegment :: NameSegment -> Name

Construct a relative name from a name segment.

O(1).

fromSegments :: NonEmpty NameSegment -> Name

Construct a relative name from a list of name segments.

>>> fromSegments ("a" :| ["b", "c"])
"a.b.c"

O(n), where n is the number of name segments.

Unsafe construction

unsafeFromString :: String -> Name

Unsafely parse a name from a string literal.

See unsafeFromText.

unsafeFromText :: HasCallStack => Text -> Name

Unsafely parse a name from a string literal.

Performs very minor validation (a name can't be empty, nor contain a # character [at least currently?]) but makes no attempt at rejecting bogus names like "foo...bar...baz".

unsafeFromVar :: Var v => v -> Name

Unsafely parse a name from a var, by first rendering the var as a string.

See unsafeFromText.

Basic queries

countSegments :: Name -> Int

Return the number of name segments in a name.

O(n), where n is the number of name segments.

isAbsolute :: Name -> Bool

Is this name absolute?

O(1).

isPrefixOf :: Name -> Name -> Bool

isPrefixOf x y returns whether x is a prefix of (or equivalent to) y, which is false if one name is relative and the other is absolute.

>>> isPrefixOf "a.b" "a.b.c"
True

>>> isPrefixOf "a.b.c" "a.b.c"
True

>>> isPrefixOf ".a.b" "a.b.c"
False

O(n), where n is the number of name segments.

endsWithReverseSegments :: Name -> [NameSegment] -> Bool

Like endsWithSegments, but accepts a list of name segments in reverse order.

Slightly more efficient than endsWithSegments.

endsWithSegments :: Name -> [NameSegment] -> Bool

endsWithSegments x y returns whether x ends with y.

>>> endsWithSegments "a.b.c" ["b", "c"]
True

>>> endsWithSegments "a.b.c" ["d"]
False

>>> endsWithSegments "a.b.c" []
True

O(n), where n is the number of name segments.

reverseSegments :: Name -> NonEmpty NameSegment

Return the name segments of a name, in reverse order.

>>> reverseSegments "a.b.c"
"c" :| ["b", "a"]

O(1).

segments :: Name -> NonEmpty NameSegment

Return the name segments of a name.

>>> segments "a.b.c"
"a" :| ["b", "c"]

O(n), where n is the number of name segments.

suffixes :: Name -> [Name]

Return all relative suffixes of a name, in descending-length order. The returned list will always be non-empty.

>>> suffixes "a.b.c"
["a.b.c", "a.b", "c"]

>>> suffixes ".a.b.c"
["a.b.c", "a.b", "c"]

Basic manipulation

makeAbsolute :: Name -> Name

Make a name absolute. No-op if the name is already absolute.

O(1).

makeRelative :: Name -> Name

Make a name relative. No-op if the name is already relative.

O(1).

parent :: Name -> Maybe Name

Compute the "parent" of a name, unless the name is only a single segment, in which case it has no parent.

>>> parent "a.b.c"
Just "b.c"

>>> parent ".a.b.c"
Just ".a.b"

>>> parent "a"
Nothing

stripNamePrefix :: Name -> Name -> Maybe Name

stripNamePrefix x y strips prefix x from name y, and returns the resulting name. Returns Nothing x is not a proper (meaning shorter-than) prefix of y.

>>> stripNamePrefix "a.b" "a.b.c"
Just "c"

>>> stripNamePrefix ".a.b" "a.b.c"
Nothing

>>> stripNamePrefix "a.b.c" "a.b.c"
Nothing

unqualified :: Name -> Name

Drop all leading segments from a name, retaining only the last segment as a relative name.

>>> unqualified "a.b.c"
"c"

>>> unqualified ".a.b.c"
"c"

To organize later

sortNames :: [Name] -> [Name]

sortNamed :: (a -> Name) -> [a] -> [a]

sortByText :: (a -> Text) -> [a] -> [a]

searchBySuffix :: Ord r => Name -> Relation Name r -> Set r

suffixFrom :: Name -> Name -> Maybe Name

shortestUniqueSuffix :: forall r. Ord r => Name -> r -> Relation Name r -> Name

toString :: Name -> String

Convert a name to a string representation.

toText :: Name -> Text

Convert a name to a string representation.

toVar :: Var v => Name -> v

Convert a name to a string representation, then parse that as a var.

splits :: HasCallStack => Name -> [([NameSegment], Name)]

Return all "splits" of a relative name, which pair a possibly-empty prefix of name segments with a suffix, such that the original name is equivalent to prefix + suffix.

Note: always returns a non-empty list, but (currently) does not use NonEmpty for convenience, as none of the call-sites care if the list is empty or not.

> splits foo.bar.baz

  prefix    suffix
  ------    ------
  ∅         foo.bar.baz
  foo       bar.baz
  foo.bar   baz

Precondition: the name is relative.

Re-exports

class Eq n => Alphabetical n where

Methods

compareAlphabetical :: n -> n -> Ordering

Exported for testing

compareSuffix :: Name -> Name -> Ordering

compareSuffix x y compares the suffix of y (in reverse segment order) that is as long as x to x (in reverse segment order).

>>> compareSuffix "b.c" "a.b.c"
EQ -- because [c,b] == [c,b]

>>> compareSuffix "b.c" "a.b.b"
LT -- because [b,b] < [c,b]

>>> compareSuffix "a.b.c" "b.c"
LT -- because [c,b] < [c,b,a]

>>> compareSuffix "b.b" "a.b.c"
GT -- because [c,b] > [b,b]

Used for suffix-based lookup of a name. For instance, given a r : Relation Name x, Relation.searchDom (compareSuffix "foo.bar") r will find all r whose name has foo.bar as a suffix.

This is only exported for testing; use searchBySuffix or shortestUniqueSuffix instead.

O(n), where n is the number of name segments.