Unison.Name.Internal

Description

The private Unison.Name innards. Prefer importing Unison.Name instead, unless you need the data constructor of Name.

Documentation

data Name

A name is an absolute-or-relative non-empty list of name segments. It is used to represent the path to a definition.

A few example names:

"foo.bar" --> Name Relative ("bar" :| ["foo"])

".foo.bar" --> Name Absolute ("bar" :| ["foo"])

"|>.-- Name Relative (":| ["|"])

"." --> Name Relative ("." :| [])

".." --> Name Absolute (".." :| [])

Constructors

Name

Fields

Position

whether the name is positioned absolutely (to some arbitrary root namespace), or relatively

(NonEmpty NameSegment)

the name segments in reverse order

isAbsolute :: Name -> Bool

Is this name absolute?

O(1).

segments :: Name -> NonEmpty NameSegment

Return the name segments of a name.

>>> segments "a.b.c"
"a" :| ["b", "c"]

O(n), where n is the number of name segments.