Unison.Names

Documentation

data Names

Constructors

Names

Fields

terms :: Relation Name Referent

types :: Relation Name TypeReference

addTerm :: Name -> Referent -> Names -> Names

addType :: Name -> TypeReference -> Names -> Names

labeledReferences :: Names -> Set LabeledDependency

Collect all references in the given Names, tagged with their type.

conflicts :: Names -> Names

filters out everything from the domain except what's conflicted

contains :: Names -> Reference -> Bool

difference :: Names -> Names -> Names

filter :: (Name -> Bool) -> Names -> Names

filterByHQs :: Set (HashQualified Name) -> Names -> Names

filterBySHs :: Set ShortHash -> Names -> Names

filterTypes :: (Name -> Bool) -> Names -> Names

map :: (Name -> Name) -> Names -> Names

makeAbsolute :: Names -> Names

makeRelative :: Names -> Names

fuzzyFind :: (Name -> Text) -> [String] -> Names -> [(Alignment, Name, Set (Either Referent TypeReference))]

hqName :: Names -> Name -> Either TypeReference Referent -> HashQualified Name

Like hqTermName and hqTypeName, but considers term and type names to conflict with each other (so will hash-qualify if there is e.g. both a term and a type named "foo").

This is useful in contexts such as printing branch diffs. Example:

Deletes:

foo foo

We want to append the hash regardless of whether or not one is a term and the other is a type.

hqTermName :: Int -> Names -> Name -> Referent -> HashQualified Name

hqTypeName :: Int -> Names -> Name -> TypeReference -> HashQualified Name

hqTermName' :: Int -> Name -> Referent -> HashQualified Name

hqTypeName' :: Int -> Name -> TypeReference -> HashQualified Name

_hqTermName :: Names -> Name -> Referent -> HashQualified Name

_hqTypeName :: Names -> Name -> TypeReference -> HashQualified Name

_hqTermAliases :: Names -> Name -> Referent -> Set (HashQualified Name)

_hqTypeAliases :: Names -> Name -> TypeReference -> Set (HashQualified Name)

mapNames :: (Name -> Name) -> Names -> Names

Map over each name in a Names.

prefix0 :: Name -> Names -> Names

prefix0 n ns prepends n to each name in ns.

Precondition: every name in ns is relative.

restrictReferences :: Set Reference -> Names -> Names

refTermsNamed :: Names -> Name -> Set TermReference

Get all terms with a specific name.

refTermsHQNamed :: Names -> HashQualified Name -> Set TermReference

Get all terms with a specific hash-qualified name.

referenceIds :: Names -> Set Id

Get all (untagged) term/type references ids in a Names.

termReferences :: Names -> Set TermReference

Returns all constructor term references. Constructors are omitted.

termReferents :: Names -> Set Referent

typeReferences :: Names -> Set TypeReference

termsNamed :: Names -> Name -> Set Referent

typesNamed :: Names -> Name -> Set TypeReference

unionLeft :: Names -> Names -> Names

unionLeftName :: Names -> Names -> Names

Guide to unionLeft* Is it ok to create new aliases for parsing? Sure.

Is it ok to create name conflicts for parsing? It's okay but not great. The user will have to hash-qualify to disambiguate.

Is it ok to create new aliases for pretty-printing? Not helpful, we need to choose a name to show. We'll just have to choose one at random if there are aliases. Is it ok to create name conflicts for pretty-printing? Still okay but not great. The pretty-printer will have to hash-qualify to disambiguate.

Thus, for parsing: unionLeftName is good if the name n on the left is the only n the user will want to reference. It allows the rhs to add aliases. unionLeftRef allows new conflicts but no new aliases. Lame? (<>) is ok for parsing if we expect to add some conflicted names, e.g. from history

For pretty-printing: Probably don't want to add new aliases, unless we don't know which Names is higher priority. So if we do have a preferred Names, don't use unionLeftName or (<>). You don't want to create new conflicts either if you have a preferred Names. So in this case, don't use unionLeftRef either. I guess that leaves unionLeft.

Not sure if the above is helpful or correct!

unionLeftRef :: Names -> Names -> Names

namesForReference :: Names -> TypeReference -> Set Name

namesForReferent :: Names -> Referent -> Set Name

shadowTerms :: [Name] -> Names -> Names

importing :: [(Name, Name)] -> Names -> Names

Given a mapping from name to qualified name, update a Names, so for instance if the input has [(Some, Optional.Some)], and Some is a constructor in the input Names, the alias Some will map to that same constructor and shadow anything else that is currently called Some.

constructorsForType :: TypeReference -> Names -> [(Name, Referent)]

expandWildcardImport :: Name -> Names -> [(Name, Name)]

Converts a wildcard import into a list of explicit imports, of the form [(suffix, full)]. Example: if io contains two functions, foo and bar, then `expandWildcardImport io` will produce `[(foo, io.foo), (bar, io.bar)]`.

isEmpty :: Names -> Bool

hashQualifyTypesRelation :: Relation Name TypeReference -> Relation (HashQualified Name) TypeReference

hashQualifyTermsRelation :: Relation Name Referent -> Relation (HashQualified Name) Referent

fromTermsAndTypes :: [(Name, Referent)] -> [(Name, TypeReference)] -> Names