Unison.Names.Scoped

Documentation

data ScopedNames

Contains all useful permutations of names scoped to a given branch.

Constructors

ScopedNames

Fields

absoluteExternalNames :: Names

relativeScopedNames :: Names

absoluteRootNames :: Names

namesAtPath :: ScopedNames -> Names

Return all names contained in the path, relative to that path.

parseNames :: ScopedNames -> Names

Includes ALL absolute names AND includes relative names for anything in the path.

prettyNames :: ScopedNames -> Names

Includes includes relative names for anything in the path, and absolute names for everything else.