Unison.NamesWithHistory

Documentation

diff :: Names -> Names -> Diff

push :: Names -> Names -> Names

lookupHQType :: SearchType -> HashQualified Name -> Names -> Set TypeReference

Find all types whose name has a suffix matching the provided HashQualified.

lookupHQType' :: SearchType -> HashQualified Name -> Names -> Set TypeReference

Find all types whose name has a suffix matching the provided HashQualified'. See lookupHQType.

lookupHQTerm :: SearchType -> HashQualified Name -> Names -> Set Referent

Find all terms whose name has a suffix matching the provided HashQualified.

lookupHQTerm' :: SearchType -> HashQualified Name -> Names -> Set Referent

Find all terms whose name has a suffix matching the provided HashQualified'. See lookupHQTerm.

lookupRelativeHQType :: SearchType -> HashQualified Name -> Names -> Set TypeReference

lookupRelativeHQType' :: SearchType -> HashQualified Name -> Names -> Set TypeReference

lookupRelativeHQTerm :: SearchType -> HashQualified Name -> Names -> Set Referent

lookupRelativeHQTerm' :: SearchType -> HashQualified Name -> Names -> Set Referent

hasTermNamed :: SearchType -> Name -> Names -> Bool

hasTypeNamed :: SearchType -> Name -> Names -> Bool

typeName :: Int -> Reference -> Names -> Set (HashQualified Name)

termNamesByLength :: Int -> Referent -> Names -> [HashQualified Name]

longestTermName :: Int -> Referent -> Names -> HashQualified Name

termName :: Int -> Referent -> Names -> Set (HashQualified Name)

lookupHQPattern :: SearchType -> HashQualified Name -> ConstructorType -> Names -> Set ConstructorReference

data Diff

Constructors

Diff

Fields

originalNames :: Names

addedNames :: Names

removedNames :: Names

data SearchType

Whether to search for exact matches or to find definitions by a suffix of their name.

Constructors

IncludeSuffixes

ExactName