Unison.Paths

Documentation

type Type v = Type v ()

type Term v = Term v ()

data Target v

Constructors

Term (Term v)

Type (Type v)

Var v

Declaration v (Term v)

vmap :: Ord v2 => (v -> v2) -> Target v -> Target v2

data PathElement

Constructors

Fn
Points at function in a function/type application

Arg
Points at the argument of a function/type application

Body
Points at the body of a lambda, let, binding, forall, or annotation

Bound
Points at the symbol bound by a `let`, lambda or forall binder

Binding !Int
Points at a particular binding in a let

Index !Int
Points at the index of a vector

Annotation
Points into the annotation

Input
Points at the left of an Arrow

Output
Points at the right of an Arrow

focus1 :: Var v => PathElement -> Path (Target v) (Target (V v)) (Target v) (Target (V v)) [v]

type Path = [PathElement]

focus :: Var v => Path -> Target v -> Maybe (Target v, Target (V v) -> Maybe (Target v), [v])

at :: Var v => Path -> Target v -> Maybe (Target v)

atTerm :: Var v => Path -> Term v -> Maybe (Term v)

atType :: Var v => Path -> Type v -> Maybe (Type v)

modify :: Var v => (Target v -> Target (V v)) -> Path -> Target v -> Maybe (Target v)

modifyTerm :: Var v => (Term v -> Term (V v)) -> Path -> Term v -> Maybe (Term v)

modifyTerm' :: Var v => (Term v -> Term (V v)) -> Path -> Term v -> Term v

modifyType :: Var v => (Type v -> Type (V v)) -> Path -> Type v -> Maybe (Type v)

inScopeAt :: Var v => Path -> Target v -> [v]

inScopeAtTerm :: Var v => Path -> Term v -> [v]

inScopeAtType :: Var v => Path -> Type v -> [v]

insertTerm :: Var v => Path -> Term v -> Maybe (Term v)

pathPrefixes :: Path -> [Path]

Return the list of all prefixes of the input path

pathExtend :: PathElement -> Path -> Path

Add an element onto the end of this Path

parent :: Path -> Maybe Path

parent' :: Path -> Path

asTerm :: Target v -> Maybe (Term v)

asType :: Target v -> Maybe (Type v)

asVar :: Target v -> Maybe v

asDeclaration :: Target v -> Maybe (v, Term v)