Unison.Project

Description

Projects.

The syntax-related parsing code (what makes a valid project name, etc) could conceivably be moved into a different package, but for now we have just defined the one blessed project/branch name syntax that we allow.

Documentation

data ProjectName

The name of a project.

projectNameUserSlug :: ProjectName -> Maybe Text

Get the user slug at the beginning of a project name, if there is one.

>>> projectNameUserSlug "@arya/lens"
Just "arya"

>>> projectNameUserSlug "lens"
Nothing

prependUserSlugToProjectName :: Text -> ProjectName -> ProjectName

Prepend a user slug to a project name, if it doesn't already have one.

>>> prependUserSlugToProjectName "arya" "lens"
"@arya/lens"

>>> prependUserSlugToProjectName "runar" "@unison/base"
"@unison/base"

>>> prependUserSlugToProjectName "???invalid???" "@unison/base"
"@unison/base"

data ProjectBranchName

The name of a branch of a project.

projectBranchNameUserSlug :: ProjectBranchName -> Maybe Text

Get the user slug at the beginning of a project branch name, if there is one.

>>> projectBranchNameUserSlug "@arya/topic"
Just "arya"

>>> projectBranchNameUserSlug "topic"
Nothing

prependUserSlugToProjectBranchName :: Text -> ProjectBranchName -> ProjectBranchName

Prepend a user slug to a project branch name, if it doesn't already have one.

>>> prependUserSlugToProjectBranchName "arya" "topic"
"@arya/topic"

>>> prependUserSlugToProjectBranchName "runar" "@unison/main"
"@unison/main"

>>> prependUserSlugToProjectBranchName "???invalid???" "@unison/main"
"@unison/main"

data ProjectAndBranch a b

A generic data structure that contains information about a project and a branch in that project.

Constructors

ProjectAndBranch

Fields

project :: a

branch :: b

projectAndBranchNamesParser :: Parsec Void Text (These ProjectName ProjectBranchName)

Orphan instances

From ProjectName Text

Instance details

Methods

from :: ProjectName -> Text

From ProjectBranchName Text

Instance details

Methods

from :: ProjectBranchName -> Text

TryFrom Text ProjectName

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text ProjectName) ProjectName

TryFrom Text ProjectBranchName

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text ProjectBranchName) ProjectBranchName

TryFrom Text (These ProjectName ProjectBranchName)

Instance details

Methods

tryFrom :: Text -> Either (TryFromException Text (These ProjectName ProjectBranchName)) (These ProjectName ProjectBranchName)

From (These ProjectName ProjectBranchName) Text
project/branch syntax for project+branch pair, with up to one side optional. Missing value means "the current one".

Instance details

Methods

from :: These ProjectName ProjectBranchName -> Text