Unison.Reference

Documentation

type Reference = Reference' Text Hash

This is the canonical representation of Reference

data Reference' t h

Either a builtin or a user defined (hashed) top-level declaration. Used for both terms and types.

Constructors

ReferenceBuiltin t

ReferenceDerived (Id' h)

Bundled Patterns

pattern Builtin :: t -> Reference' t h

pattern DerivedId :: Id' h -> Reference' t h

pattern Derived :: h -> Pos -> Reference' t h

_DerivedId :: Prism' Reference Id

type Id = Id' Hash

data Id' h

Pos is a position into a cycle, as cycles are hashed together.

Constructors

Id h Pos

type Pos = Word64

type CycleSize = Word64

type Size = CycleSize

type TermReference = Reference

A term reference.

type TermReferenceId = Id

A term reference id.

type TypeReference = Reference

A type declaration reference.

type TypeReferenceId = Id

A type declaration reference id.

derivedBase32Hex :: Text -> Pos -> Maybe Reference

component :: Hash -> [k] -> [(k, Id)]

components :: [(Hash, [k])] -> [(k, Id)]

groupByComponent :: [(k, Reference)] -> [[(k, Reference)]]

componentFor :: Hash -> [a] -> [(Id, a)]

componentFromLength :: Hash -> CycleSize -> Set Id

unsafeFromText :: Text -> Reference

isPrefixOf :: ShortHash -> Reference -> Bool

fromText :: Text -> Either String Reference

todo: take a (Reference -> CycleSize) so that readSuffix doesn't have to parse the size from the text. examples:

builtins don’t have cycles >>> fromText "##Text.take" Right ##Text.take

derived, no cycle >>> fromText "#dqp2oi4iderlrgp2h11sgkff6drk92omo4c84dncfhg9o0jn21cli4lhga72vlchmrb2jk0b3bdc2gie1l06sqdli8ego4q0akm3au8" Right #dqp2o

derived, part of cycle >>> fromText "#dqp2oi4iderlrgp2h11sgkff6drk92omo4c84dncfhg9o0jn21cli4lhga72vlchmrb2jk0b3bdc2gie1l06sqdli8ego4q0akm3au8.12345" Right #dqp2o.12345

Errors with Left on invalid hashes >>> fromText "#invalid_hash.12345" Left "Invalid hash: "invalid_hash""

readSuffix :: Text -> Either String Pos

showShort :: Int -> Reference -> Text

showSuffix :: Pos -> Text

toHash :: Reference -> Maybe Hash

toId :: Reference -> Maybe Id

fromId :: Id -> Reference

toText :: Reference -> Text

idToText :: Id -> Text

unsafeId :: Reference -> Id

toShortHash :: Reference -> ShortHash

idToHash :: Id -> Hash

idToShortHash :: Id -> ShortHash

isBuiltin :: Reference -> Bool