Unison.Util.Nametree

Nametree

data Nametree a

A nametree has a value, and a collection of children nametrees keyed by name segment.

Constructors

Nametree

Fields

value :: !a

children :: !(Map NameSegment (Nametree a))

traverseNametreeWithName :: Applicative f => ([NameSegment] -> a -> f b) -> Nametree a -> f (Nametree b)

Traverse over a nametree, with access to the list of name segments (in reverse order) leading to each value.

unfoldNametree :: (a -> (b, Map NameSegment a)) -> a -> Nametree b

Build a nametree from a seed value.

Flattening and unflattening

flattenNametree :: forall a b. Ord b => (a -> Map NameSegment b) -> Nametree a -> BiMultimap b Name

flattenNametree organizes a nametree like

"foo" = #foo
"foo": {
  "bar" = #bar
  "bar": {
    "baz" = #baz
  }
}

into an equivalent-but-flat association between names and definitions, like

{
  "foo" = #bar,
  "foo.bar" = #bar,
  "foo.bar.baz" = #baz
}

flattenNametrees :: (Ord term, Ord typ) => Nametree (DefnsF (Map NameSegment) term typ) -> Defns (BiMultimap term Name) (BiMultimap typ Name)

Like flattenNametree, but works on both the types and terms namespace at once.

unflattenNametree :: Ord a => Map Name a -> Nametree (Map NameSegment a)

unflattenNametree organizes an association between names and definitions like

{
  "foo" = #bar,
  "foo.bar" = #bar,
  "foo.bar.baz" = #baz
}

into an equivalent-but-less-flat nametree, like

"foo" = #foo
"foo": {
  "bar" = #bar
  "bar": {
    "baz" = #baz
  }
}

unflattenNametrees :: (Ord term, Ord typ) => DefnsF (Map Name) term typ -> Nametree (DefnsF (Map NameSegment) term typ)

Like unflattenNametree, but works on both the types and terms namespace at once.