Unison.Var

Documentation

class (Show v, Var v) => Var v where

A class for variables. Variables may have auxiliary information which may not form part of their identity according to Eq / Ord. Laws:

`typeOf (typed n) == n`

`typeOf (ABT.freshIn vs v) == typeOf v`: freshIn does not alter the name

Methods

typed :: Type -> v

typeOf :: v -> Type

freshId :: v -> Word64

freshenId :: Word64 -> v -> v

data Type

Constructors

User Text

Inference InferenceType

MissingResult

Blank

UnnamedReference Id
An unnamed reference, created during unhashing a term/decl component.

UnnamedWatch WatchKind Text

Eta

ANFBlank

Float

Pattern

Irrelevant

Delay

data InferenceType

Constructors

Ability

Input

Output

PatternPureE

PatternPureV

PatternBindE

PatternBindV

TypeConstructor

TypeConstructorArg

Other

bakeId :: Var v => v -> v

blank :: Var v => v

freshIn :: Var v => Set v -> v -> v

inferAbility :: Var v => v

inferInput :: Var v => v

inferOther :: Var v => v

inferOutput :: Var v => v

inferPatternBindE :: Var v => v

inferPatternBindV :: Var v => v

inferPatternPureE :: Var v => v

inferPatternPureV :: Var v => v

inferTypeConstructor :: Var v => v

inferTypeConstructorArg :: Var v => v

isAction :: Var v => v -> Bool

Currently, actions in blocks are encoded as bindings with names of the form _123 (an underscore, followed by 1 or more digits). This function returns True if the input variable has this form.

Various places check for this (the pretty-printer, to determine how to print the binding), and the typechecker (to decide if it should ensure the binding has type ()).

missingResult :: Var v => v

name :: Var v => v -> Text

nameStr :: Var v => v -> String

named :: Var v => Text -> v

nameds :: Var v => String -> v

rawName :: Type -> Text

reset :: Var v => v -> v

uncapitalize :: Var v => v -> v

universallyQuantifyIfFree :: forall v. Var v => v -> Bool

unnamedRef :: Var v => Id -> v

unnamedTest :: Var v => Text -> v