Unison.ABT

Description

Based on: http://semantic-domain.blogspot.com/2015/03/abstract-binding-trees.html

Types

data ABT (f :: Type -> Type) v r

Constructors

Var v

Cycle r

Abs v r

Tm (f r)

data Term (f :: Type -> Type) v a

At each level in the tree, we store the set of free variables and a value of type a. Variables are of type v.

Constructors

Term

Fields

freeVars :: Set v

annotation :: a

out :: ABT f v (Term f v a)

data Term' (f :: Type -> Type) v a x

Constructors

Term'

Fields

freeVars' :: Set v

annotation' :: a

out' :: ABT f v x

class Ord v => Var v where

A class for avoiding accidental variable capture

`Set.notMember (freshIn vs v) vs`: freshIn returns a variable not used in the Set

Methods

freshIn :: Set v -> v -> v

data V v

Constructors

Free v

Bound v

data Subst f v a

Constructors

Subst

Fields

freshen :: forall m v'. Monad m => (v -> m v') -> m v'

bind :: Term f v a -> Term f v a

bindInheritAnnotation :: forall b. Term f v b -> Term f v a

variable :: v

Combinators & Traversals

fresh :: Var v => Term f v a -> v -> v

unvar :: V v -> v

freshenS :: (Var v, MonadState (Set v) m) => v -> m v

Freshens the given variable wrt. the set of used variables tracked by state. Adds the result to the set of used variables.

freshInBoth :: forall v (f :: Type -> Type) a. Var v => Term f v a -> Term f v a -> v -> v

Produce a variable which is free in both terms

freshenBothWrt :: (Var v, Traversable f) => Term f v a -> Term f v a -> Term f v a -> (Term f v a, Term f v a)

freshenWrt :: (Var v, Traversable f) => (v -> v) -> Term f v a -> [Term f v a] -> [Term f v a]

visit :: forall (f :: Type -> Type) g v a. (Traversable f, Applicative g, Ord v) => (Term f v a -> Maybe (g (Term f v a))) -> Term f v a -> g (Term f v a)

`visit f t` applies an effectful function to each subtree of t and sequences the results. When f returns Nothing, visit descends into the children of the current subtree. When f returns `Just t2`, visit replaces the current subtree with t2. Thus: `visit (const Nothing) t == pure t` and `visit (const (Just (pure t2))) t == pure t2`

visit' :: (Traversable f, Monad g, Ord v) => (f (Term f v a) -> g (f (Term f v a))) -> Term f v a -> g (Term f v a)

Apply an effectful function to an ABT tree top down, sequencing the results.

visit_ :: (Traversable f, Applicative g, Ord v) => (f (Term f v a) -> g ()) -> Term f v a -> g (Term f v a)

Apply an effectful function to an ABT tree top down, sequencing the results.

visitPure :: forall (f :: Type -> Type) v a. (Traversable f, Ord v) => (Term f v a -> Maybe (Term f v a)) -> Term f v a -> Term f v a

visit specialized to the Identity effect.

changeVars :: (Foldable f, Functor f, Var v) => Map v v -> Term f v a -> Term f v a

allVars :: forall (f :: Type -> Type) v a. Foldable f => Term f v a -> [v]

numberedFreeVars :: (Ord v, Foldable f) => Term f v a -> Map v Int

subterms :: forall v (f :: Type -> Type) a. (Ord v, Traversable f) => Term f v a -> [Term f v a]

annotateBound :: (Ord v, Foldable f, Functor f) => Term f v a -> Term f v (a, Set v)

rebuildUp :: (Ord v, Foldable f, Functor f) => (f (Term f v a) -> f (Term f v a)) -> Term f v a -> Term f v a

rebuildMaybeUp :: (Ord v, Foldable f, Functor f) => (Term f v a -> Maybe (Term f v a)) -> Term f v a -> Maybe (Term f v a)

rebuildUp' :: (Ord v, Foldable f, Functor f) => (Term f v a -> Term f v a) -> Term f v a -> Term f v a

reannotateUp :: (Ord v, Foldable f, Functor f, Monoid b) => (Term f v a -> b) -> Term f v a -> Term f v (a, b)

rewriteDown :: (Traversable f, Ord v) => (Term f v a -> Term f v a) -> Term f v a -> Term f v a

transform :: (Ord v, Foldable g, Functor g) => (forall a1. f a1 -> g a1) -> Term f v a -> Term g v a

transformM :: (Ord v, Monad m, Traversable g) => (forall a1. f a1 -> m (g a1)) -> Term f v a -> m (Term g v a)

foreachSubterm :: forall (f :: Type -> Type) g v a b. (Traversable f, Applicative g) => (Term f v a -> g b) -> Term f v a -> g [b]

freeVarOccurrences :: (Traversable f, Ord v) => Set v -> Term f v a -> [(v, a)]

isFreeIn :: Ord v => v -> Term f v a -> Bool

True if v is a member of the set of free variables of t

occurrences :: (Foldable f, Var v) => v -> Term f v a -> Int

extraMap :: Functor g => (forall k. f k -> g k) -> Term f v a -> Term g v a

vmap :: forall (f :: Type -> Type) v' v a. (Functor f, Foldable f, Ord v') => (v -> v') -> Term f v a -> Term f v' a

vmapM :: forall m (f :: Type -> Type) v2 v a. (Applicative m, Traversable f, Foldable f, Ord v2) => (v -> m v2) -> Term f v a -> m (Term f v2 a)

amap :: (Functor f, Foldable f, Ord v) => (a -> a2) -> Term f v a -> Term f v a2

rename :: forall (f :: Type -> Type) v a. (Foldable f, Functor f, Var v) => v -> v -> Term f v a -> Term f v a

renames old to new in the given term, ignoring subtrees that bind old

renames :: (Foldable f, Functor f, Var v) => Map v v -> Term f v a -> Term f v a

subst :: (Foldable f, Functor f, Var v) => v -> Term f v a -> Term f v a -> Term f v a

`subst v e body` substitutes e for v in body, avoiding capture by renaming abstractions in body

substs :: (Foldable f, Functor f, Var v) => [(v, Term f v a)] -> Term f v a -> Term f v a

`substs [(t1,v1), (t2,v2), ...] body` performs multiple simultaneous substitutions, avoiding capture

substInheritAnnotation :: forall (f :: Type -> Type) v b a. (Foldable f, Functor f, Var v) => v -> Term f v b -> Term f v a -> Term f v a

substsInheritAnnotation :: forall (f :: Type -> Type) v b a. (Foldable f, Functor f, Var v) => [(v, Term f v b)] -> Term f v a -> Term f v a

find :: (Ord v, Foldable f, Functor f) => (Term f v a -> FindAction x) -> Term f v a -> [x]

find' :: (Ord v, Foldable f, Functor f) => (Term f v a -> Bool) -> Term f v a -> [Term f v a]

data FindAction x

Constructors

Found x

Prune

Continue

containsExpression :: forall f v a. (Var v, forall a. Eq a => Eq (f a), Traversable f) => Term f v a -> Term f v a -> Bool

Core logic of structured find. Works for any base functor. Returns True if there's a subexpression of tm which matches query0 for some assignment of variables.

rewriteExpression :: forall f v a. (Var v, Show v, forall a. Eq a => Eq (f a), forall a. Show a => Show (f a), Traversable f) => Term f v a -> Term f v a -> Term f v a -> Maybe (Term f v a)

Core logic of structured find and replace. Works for any base functor. Returns Nothing if no replacements.

Optics

baseFunctor_ :: Applicative m => (f (Term f v a) -> m (f (Term f v a))) -> Term f v a -> m (Term f v a)

rewriteDown_ :: (Traversable f, Monad m, Ord v) => (Term f v a -> m (Term f v a)) -> Term f v a -> m (Term f v a)

Setter' (Term f v a) (Term f v a)

Safe Term constructors & Patterns

annotate :: a -> Term f v a -> Term f v a

Replace the annotation with the given argument.

annotatedVar :: a -> v -> Term f v a

var :: v -> Term f v ()

tm :: (Foldable f, Ord v) => f (Term f v ()) -> Term f v ()

tm' :: (Foldable f, Ord v) => a -> f (Term f v a) -> Term f v a

abs :: Ord v => v -> Term f v () -> Term f v ()

absChain :: Ord v => [v] -> Term f v () -> Term f v ()

absChain' :: Ord v => [(a, v)] -> Term f v a -> Term f v a

abs' :: Ord v => a -> v -> Term f v a -> Term f v a

absr :: (Functor f, Foldable f, Var v) => v -> Term f (V v) () -> Term f (V v) ()

unabs :: forall (f :: Type -> Type) v a. Term f v a -> ([v], Term f v a)

unabsA :: Term f v a -> ([(a, v)], Term f v a)

dropAbs :: Int -> Term f v a -> Term f v a

cycle :: Term f v () -> Term f v ()

cycle' :: a -> Term f v a -> Term f v a

cycler :: (Functor f, Foldable f, Var v) => [v] -> Term f (V v) () -> Term f (V v) ()

pattern Abs' :: (Foldable f, Functor f, Var v) => Subst f v a -> Term f v a

pattern Abs'' :: v -> Term f v a -> Term f v a

pattern AbsN' :: [v] -> Term f v a -> Term f v a

pattern AbsNA' :: [(a, v)] -> Term f v a -> Term f v a

pattern Var' :: v -> Term f v a

pattern Cycle' :: [v] -> f (Term f v a) -> Term f v a

pattern Cycle'' :: Term f v a -> Term f v a

pattern CycleA' :: a -> [(a, v)] -> Term f v a -> Term f v a

pattern Tm' :: f (Term f v a) -> Term f v a

Algorithms

components :: Var v => [(v, Term f v a)] -> [[(v, Term f v a)]]

orderedComponents :: Var v => [(v, Term f v a)] -> [[(v, Term f v a)]]