Unison.ConstructorReference

Description

The constructor reference type.

Documentation

data GConstructorReference r

A reference to a constructor is represented by a reference to its type declaration, plus the ordinal constructor id.

Constructors

ConstructorReference !r !ConstructorId

type ConstructorReference = GConstructorReference TypeReference

type ConstructorReferenceId = GConstructorReference TypeReferenceId

reference_ :: Lens (GConstructorReference r) (GConstructorReference s) r s

A lens onto the reference part of a constructor reference.

toId :: ConstructorReference -> Maybe ConstructorReferenceId

toShortHash :: ConstructorReference -> ShortHash